Turn frames into package
